Moved hypercall parameters into separate module
