+++ uhyve-interface/src/parameters.rs
@@ -1,0 +1,109 @@
+//! Parameters for hypercalls.
+
+use crate::MAX_ARGC_ENVC;
+use x86_64::PhysAddr;
+
+/// Parameters for a [`Cmdsize`](Hypercall::Cmdsize) hypercall which provides the lengths of the items in the argument end environment vector.
+/// Parameters for a [`Cmdsize`](Hypercall::Cmdsize) hypercall which provides the lengths of the items in the argument end environment vector.
+#[repr(C, packed)]
+#[derive(Debug, Copy, Clone)]
+pub struct SysCmdsize {
+	/// Nr of items in the kernel command line.
+	pub argc: i32,
+	/// Lengths of the items in the kernel command line.
+	pub argsz: [i32; MAX_ARGC_ENVC],
+	/// Nr of items in the environment.
+	pub envc: i32,
+	/// Length of the items in the environment.
+	pub envsz: [i32; MAX_ARGC_ENVC],
+}
+
+/// Parameters for a [`Cmdval`](Hypercall::Cmdval) hypercall, which copies the arguments end environment of the application into the VM's memory.
+#[repr(C, packed)]
+#[derive(Debug, Copy, Clone)]
+pub struct SysCmdval {
+	/// Pointer to a memory section in the VM memory large enough to store the argument string.
+	pub argv: PhysAddr,
+	/// Pointer to a memory section in the VM memory large enough to store the environment values.
+	pub envp: PhysAddr,
+}
+
+/// Parameters for a [`Exit`](Hypercall::Exit) hypercall.
+#[repr(C, packed)]
+#[derive(Debug, Copy, Clone)]
+pub struct SysExit {
+	/// The return code of the guest.
+	pub arg: i32,
+}
+
+/// Parameters for a [`FileUnlink`](Hypercall::FileUnlink) hypercall.
+#[repr(C, packed)]
+#[derive(Debug, Copy, Clone)]
+pub struct SysUnlink {
+	/// Address of the file that should be unlinked.
+	pub name: PhysAddr,
+	/// On success, `0` is returned.  On error, `-1` is returned.
+	pub ret: i32,
+}
+
+/// Parameters for a [`FileWrite`](Hypercall::FileWrite) hypercall.
+#[repr(C, packed)]
+#[derive(Debug, Copy, Clone)]
+pub struct SysWrite {
+	/// File descriptor of the file.
+	pub fd: i32,
+	/// Buffer to be written into the file.
+	pub buf: PhysAddr,
+	/// Number of bytes in the buffer to be written.
+	pub len: usize,
+}
+
+/// Parameters for a [`FileRead`](Hypercall::FileRead) hypercall.
+#[repr(C, packed)]
+#[derive(Debug, Copy, Clone)]
+pub struct SysRead {
+	/// File descriptor of the file.
+	pub fd: i32,
+	/// Buffer to read the file into.
+	pub buf: PhysAddr,
+	/// Number of bytes to read into the buffer.
+	pub len: usize,
+	/// Number of bytes read on success. `-1` on failure.
+	pub ret: isize,
+}
+
+/// Parameters for a [`FileClose`](Hypercall::FileClose) hypercall
+#[repr(C, packed)]
+#[derive(Debug, Copy, Clone)]
+pub struct SysClose {
+	/// File descriptor of the file.
+	pub fd: i32,
+	/// Zero on success, `-1` on failure.
+	pub ret: i32,
+}
+
+/// Parameters for a [`FileOpen`](Hypercall::FileOpen) hypercall
+#[repr(C, packed)]
+#[derive(Debug, Copy, Clone)]
+pub struct SysOpen {
+	/// Pathname of the file to be opened.
+	pub name: PhysAddr,
+	/// Posix file access mode flags.
+	pub flags: i32,
+	/// Access permissions upon opening/creating a file.
+	pub mode: i32,
+	/// File descriptor upon successful opening or `-1` upon failure.
+	pub ret: i32,
+}
+
+/// Parameters for a [`FileLseek`](Hypercall::FileLseek) hypercall
+#[repr(C, packed)]
+#[derive(Debug, Copy, Clone)]
+pub struct SysLseek {
+	/// File descriptor of the file.
+	pub fd: i32,
+	/// Offset in the file.
+	pub offset: isize,
+	/// `whence` value of the lseek call.
+	pub whence: i32,
+}
